rename dickcord.rs to dickcord_old for refactoring
